Now launching the update Activity.Had to move the even listener to the same class and put the methods that calls the UpdateActivity outside of it.
